Refactoring: extract icon functionality into document base.
By extracting the icon-by-mimetype functionality from the document FTI
class into the BaseDocumentMixin, we make it possible to reuse it in
other documentish types.

Because of the inheritance order we still need to define "getIcon" and
"icon" at the FTI class.

This change adds new methods to the mail FTI class, but these methods
are not used. The getIcon method of the mail FTI is not changed.
