+++ opengever/document/document.py
@@ -19,8 +19,6 @@ from plone.dexterity.content import Item
 from plone.directives import form
 from plone.namedfile import field
 from plone.namedfile.file import NamedBlobFile
-from Products.CMFCore.utils import getToolByName
-from Products.MimetypesRegistry.common import MimeTypeException
 from z3c.form import validator
 from zope import schema
 from zope.component import getMultiAdapter
@@ -138,10 +136,6 @@ class Document(Item, BaseDocumentMixin):
 
         return False
 
-    def surrender(self, relative_to_portal=1):
-        return super(Document, self).getIcon(
-            relative_to_portal=relative_to_portal)
-
     def related_items(self):
         relations = IRelatedDocuments(self).relatedItems
         if relations:
@@ -149,54 +143,12 @@ class Document(Item, BaseDocumentMixin):
         return []
 
     def getIcon(self, relative_to_portal=1):
-        """Calculate the icon using the mime type of the file
-        """
-        utool = getToolByName(self, 'portal_url')
-
-        mimetypeitem = self.get_mimetype()
-        if not mimetypeitem:
-            return self.surrender(relative_to_portal)
-
-        icon = mimetypeitem[0].icon_path
-
-        if relative_to_portal:
-            return icon
-        else:
-            # Relative to REQUEST['BASEPATH1']
-            res = utool(relative=1) + '/' + icon
-            while res[:1] == '/':
-                res = res[1:]
-            return res
+        return self.get_mimetype_icon(relative_to_portal)
 
     def icon(self):
         """for ZMI
         """
         return self.getIcon()
-
-    def get_mimetype(self):
-        """Return the mimetype as object. If there is no matching mimetype,
-           it returns False.
-        """
-        mtr = getToolByName(self, 'mimetypes_registry', None)
-
-        field = self.file
-        if not field or not field.getSize():
-            # there is no file
-            return False
-
-        # get icon by content type
-        contenttype = field.contentType
-        mimetypeitem = None
-        try:
-            mimetypeitem = mtr.lookup(contenttype)
-        except MimeTypeException, msg:
-            LOG.error(
-                'MimeTypeException for %s. Error is: %s' % (
-                    self.absolute_url(), str(msg)))
-        if not mimetypeitem:
-            # not found
-            return False
-        return mimetypeitem
 
     def as_shadow_document(self):
         """Force a document into the shadow state.
